+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # v0.4.0 (unreleased)
 ### FEATURES
+#### Resources
+* **New Resource:** `pingfederate_oauth_client` ([#111](https://github.com/pingidentity/terraform-provider-pingfederate/pull/111))
 
 #### Data Sources
+* **New Data Source:** `pingfederate_oauth_client` ([#111](https://github.com/pingidentity/terraform-provider-pingfederate/pull/111))
 * **New Data Source:** `pingfederate_session_policies_global` ([#117](https://github.com/pingidentity/terraform-provider-pingfederate/pull/117))